Second Attempt to Enable Material for MkDocs (#272)
It turned out that the usage of Assets folder collides with the assets what Material for MkDocs expected on the letter case. And it ended up with 404 errors when reaching out to the assets/* URL. One to workaround this problem is to avoid the usage of Assets, so we won't have collisions later with the MkDocs system.